Change DataServlet into three separate servlets.
