+++ portfolio/src/main/webapp/comments.html
@@ -40,7 +40,7 @@
   <body onpageshow="getAndDisplayComments()">
     <div  id="content" class="fade-in">
 
-      <form class="center" action="/data" method="POST">
+      <form class="center" action="/new-comment" method="POST">
         Leave a comment!
         <br>
         <textarea name="comment" rows="3"></textarea>
